pyproject.toml
[tool.poetry]
name = "Strong Fitness Studio App"
version = "0.1.0"
description = "Strong Fitness Studio App"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.49.0"
pydantic-settings = "^2.8.1"
streamlit-calendar = "^1.4.0"
psycopg2-binary = "^2.9.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
